Refactor projet
To make the usage bit more straight-forward (and since we are just using express now), integrate express and the chori-express to the main package itself. This way the client only needs to import the main package and does not need import router-specifics. Later on, we can implement mechanism that allows the client to override the router or replace it completely.