Prepare for API testing
In order us to test properly, we have to allow manual registration of router integrations (was named router implementation previously, before renamed in this commit). This involves exposing new method on the application for registering router integrations as well as refactoring how the default routing port is set. Previously it was given in start command, which does not make sense since the start command needs to be called with custom routing integrations too. If the custom routing integration is used, the server startup is delegated to the user (the port is used only when the user selects default routing). Since the server startup is delegated to the user, the user must also decide the port.