Implement additional check for component scanning to determine if the scanned file is actually decorated
We use typescript AST for this. Kinda crude check, but check if the source file has __decorate function. If it does, it has decorators and we want to require it. This should later be modified to have more specific checks (to actually find out if the source file uses the polku decorators), but for now this should block unwanted requires at least to some extent.